third_party/blink/web_tests/external/wpt/loading/lazyload/original-base-url-applied-2-tentative-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Deferred images with loading='lazy' use the original base URL specified at the parse time
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a testharness.js-based test.
FAIL Deferred images with loading='lazy' use the original base URL specified at the parse time assert_unreached: The image request should not load relative to the current (incorrect) base URL. Reached unreachable code
Harness: the test ran to completion.
